pyproject.toml
[project]
name = "cartoonizer"
version = "0.1.0"
description = "A project to cartoonize images."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    {email = "you@example.com"},
    {name = "Your Name"},
]
dependencies = [
    "transformers==4.52.4",
    "diffusers==0.33.1",
    "peft==0.15.2",
    "opencv-python==4.11.0.86",
    "protobuf==5.29.4",
    "sentencepiece==0.2.0",
    "jupyter==1.1.1",
    "torchao==0.11.0",
    "boto3==1.38.37",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"